1) Изменены запросы по фильмам.
